requirements.txt
discord.py
waifuim.py
asyncio
jikanpy